global: Add character device creation
This commit implements the creation and disponibility of the gpio_ctl character device.
The device shall be used to serve as communication bridge between user and kernel address space
for the custom control of gpios.
Moreover, this commit also defines the an initial set of file operation registered for the
char device driver and also a logging macro which wraps an identifier to ease dmesg log analysis

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>