global: Add gpio descriptors initialization and release
This commit implements the gpio descriptors management (initialization and release).
The driver uses the legacy API to fetch the pins and then parses it to a descriptor for control.
The use of the legacy API just the gpio request and release is due to failure on requesting the
gpios descriptors with the new API.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>