pwm: Add pwm channels management
This commit implements an initial ioctl command to allow user-space applications
to request an active pwm channel.
When requested and the pwm device assigned, pwm states may be applied to it in order
to define period, duty-cycle and polarity.
The pwm devices are released upon driver unloading and a ioctl command shall be created
to allow the user-space application to free its requested pwm resources (sync shall be required),
though by now the module is being designed for single application use.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>