pwm: Remove user-space dependency with pwm kernel include file
To avoid the requirement of an user-space application which desires to benefit
from this module having to include the pwm kernel include file due to the pwm_state
struct declaration (avoiding any yocto workarounds to have it in the application recipe sysroot).

With this commit, a local intermediary struct is declared and used to bridge between the application and the pwm state.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>